pyproject.toml
[project]
name = "swift-scraper"
version = "0.1.0"
description = "SWIFT codes scraper mapping to n8n workflow"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pymongo[srv]>=4.6.0",
    "dnspython>=2.6.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
